Hw5 auth

+++ src/services/contacts.js
@@ -2,6 +2,7 @@ import paginationData from '../utils/calculationPaginationData.js';
 import { Contact } from '../db/model/Contact.js';
 
 export const getAllContacts = async ({
+  userId,
   page,
   perPage,
   sortBy,
@@ -9,14 +10,16 @@ export const getAllContacts = async ({
   filter = {},
 }) => {
   const skip = (page - 1) * perPage;
-  const contactQuery = Contact.find();
+  const contactQuery = Contact.find({ userId });
   if (filter.contactType) {
     contactQuery.where('contactType').equals(filter.contactType);
   }
   if (filter.isFavourite) {
     contactQuery.where('isFavourite').equals(filter.isFavourite);
   }
-  const totalItems = await Contact.find().merge(contactQuery).countDocuments();
+  const totalItems = await Contact.find({ userId })
+    .merge(contactQuery)
+    .countDocuments();
   const contactItems = await contactQuery
     .skip(skip)
     .limit(perPage)
@@ -25,13 +28,13 @@ export const getAllContacts = async ({
   const paginationInfo = paginationData(totalItems, page, perPage);
 
   return {
-    data: contactItems,
+    contacts: contactItems,
     ...paginationInfo,
   };
 };
 
-export const getContactById = async (contactId) => {
-  const contact = await Contact.findById(contactId);
+export const getContactById = async ({ contactId, userId }) => {
+  const contact = await Contact.findById({ _id: contactId, userId });
   return contact;
 };
 
@@ -40,14 +43,24 @@ export const createContact = async (payload) => {
   return contact;
 };
 
-export const patchContact = async (contactId, payload, options = {}) => {
-  const contact = await Contact.findOneAndUpdate({ _id: contactId }, payload, {
-    ...options,
-  });
+export const patchContact = async ({
+  contactId,
+  payload,
+  userId,
+  options = {},
+}) => {
+  const contact = await Contact.findOneAndUpdate(
+    { _id: contactId, userId },
+    payload,
+    {
+      new: true,
+      ...options,
+    },
+  );
   return contact;
 };
 
-export const deleteContact = async (contactId) => {
-  const contact = await Contact.findOneAndDelete({ _id: contactId });
+export const deleteContact = async (contactId, userId) => {
+  const contact = await Contact.findOneAndDelete({ _id: contactId, userId });
   return contact;
 };
